restore and move
