+++ packages/slack/src/index.ts
@@ -10,5 +10,5 @@ export const supportedGrowiCommands: string[] = [
 export * from './interfaces/growi-command';
 export * from './models/errors';
 export * from './utils/slash-command-parser';
-export * from './utils/verifying-is-slack-request';
+export * from './middlewares/verifying-is-slack-request';
 export * from './utils/block-creater';
